refactor(web): extract renderMarkdown into shared web/src/lib/markdown.ts (#125)
- Create web/src/lib/markdown.ts with exported renderMarkdown() and
  internal inlineFormat() — canonical from WikiView (more complete:
  H4 support, space-y-1 lists)
- Remove duplicate renderMarkdown+inlineFormat from WikiView.vue
- Remove duplicate renderMarkdown+inlineFormat from NotificationsView.vue
- Both views now import { renderMarkdown } from '../lib/markdown'

Minor: NotificationsView had slightly drifted CSS (bg-gray-950 code
blocks, space-y-0.5 lists) — unified to the WikiView canonical values.
No functional difference for users.

72/72 tests passing, tsc clean.
