+++ web/src/views/NotificationsView.vue
@@ -49,6 +49,7 @@
 
 <script setup lang="ts">
 import { ref, onMounted } from 'vue'
+import { renderMarkdown } from '../lib/markdown'
 import { apiFetch } from '../lib/api'
 
 interface Notification {
@@ -79,81 +80,6 @@ function toggleExpand(id: number) {
   expanded.value = next
 }
 
-function inlineFormat(text: string): string {
-  return text
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/\*\*\*(.*?)\*\*\*/g, '<strong><em>$1</em></strong>')
-    .replace(/\*\*(.*?)\*\*/g, '<strong>$1</strong>')
-    .replace(/\*(.*?)\*/g, '<em>$1</em>')
-    .replace(/`([^`]+)`/g, '<code class="bg-gray-800 px-1 rounded text-blue-300 font-mono text-xs">$1</code>')
-    .replace(/\[([^\]]+)\]\(([^)]+)\)/g, (_, linkText, url) => {
-      const safe = /^https?:\/\//i.test(url) || url.startsWith('/') || url.startsWith('#')
-      return `<a href="${safe ? url : '#'}" target="_blank" rel="noopener" class="text-blue-400 hover:underline">${linkText}</a>`
-    })
-}
-
-function renderMarkdown(md: string): string {
-  if (!md) return ''
-  const lines = md.split('\n')
-  const out: string[] = []
-  let inCode = false
-  let inList = false
-  let codeLines: string[] = []
-
-  function closeList() {
-    if (inList) { out.push('</ul>'); inList = false }
-  }
-
-  for (const line of lines) {
-    if (line.startsWith('```')) {
-      if (!inCode) {
-        closeList(); inCode = true; codeLines = []
-      } else {
-        const escaped = codeLines.join('\n').replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;')
-        out.push(`<pre class="bg-gray-950 border border-gray-800 rounded p-3 my-2 overflow-x-auto"><code class="font-mono text-xs text-green-300">${escaped}</code></pre>`)
-        inCode = false; codeLines = []
-      }
-      continue
-    }
-    if (inCode) { codeLines.push(line); continue }
-
-    const h3 = line.match(/^###\s+(.+)/)
-    const h2 = line.match(/^##\s+(.+)/)
-    const h1 = line.match(/^#\s+(.+)/)
-    if (h3) { closeList(); out.push(`<h3 class="text-base font-semibold text-gray-100 mt-4 mb-1">${inlineFormat(h3[1])}</h3>`); continue }
-    if (h2) { closeList(); out.push(`<h2 class="text-lg font-bold text-gray-100 mt-5 mb-2 pb-1 border-b border-gray-700">${inlineFormat(h2[1])}</h2>`); continue }
-    if (h1) { closeList(); out.push(`<h1 class="text-xl font-bold text-gray-100 mt-5 mb-2">${inlineFormat(h1[1])}</h1>`); continue }
-
-    if (/^---+$/.test(line.trim())) { closeList(); out.push('<hr class="border-gray-700 my-3" />'); continue }
-
-    const li = line.match(/^[-*+]\s+(.+)/)
-    if (li) {
-      if (!inList) { out.push('<ul class="list-disc list-inside my-2 space-y-0.5 pl-2">'); inList = true }
-      out.push(`<li>${inlineFormat(li[1])}</li>`)
-      continue
-    }
-
-    const oli = line.match(/^\d+\.\s+(.+)/)
-    if (oli) {
-      if (!inList) { out.push('<ol class="list-decimal list-inside my-2 space-y-0.5 pl-2">'); inList = true }
-      out.push(`<li>${inlineFormat(oli[1])}</li>`)
-      continue
-    }
-
-    if (line.trim() === '') { closeList(); out.push('<div class="my-1"></div>'); continue }
-    closeList()
-    out.push(`<p class="my-0.5">${inlineFormat(line)}</p>`)
-  }
-
-  if (inCode && codeLines.length) {
-    const escaped = codeLines.join('\n').replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;')
-    out.push(`<pre class="bg-gray-950 border border-gray-800 rounded p-3 my-2 overflow-x-auto"><code class="font-mono text-xs text-green-300">${escaped}</code></pre>`)
-  }
-  closeList()
-  return out.join('\n')
-}
 
 async function loadNotifications() {
   try {
